// model/accountsInfoRequest.ts
/**
 * Key Connect API
 * Blockchain access made simple
 *
 * The version of the OpenAPI document: 1.1.0
 * Contact: manthan@keyconnect.app
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { AccountsInfoRequestItem } from './accountsInfoRequestItem';

export class AccountsInfoRequest {
    'accounts'?: Array<AccountsInfoRequestItem>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "accounts",
            "baseName": "accounts",
            "type": "Array<AccountsInfoRequestItem>"
        }    ];

    static getAttributeTypeMap() {
        return AccountsInfoRequest.attributeTypeMap;
    }
}