// model/blockchainNetworkServerStatus.ts
/**
 * Key Connect API
 * Blockchain access made simple
 *
 * The version of the OpenAPI document: 1.1.0
 * Contact: manthan@keyconnect.app
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* Represents status of a blockchain network.
*/
export class BlockchainNetworkServerStatus {
    'host'?: string;
    'status'?: BlockchainNetworkServerStatus.StatusEnum;
    'lastCheck'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "host",
            "baseName": "host",
            "type": "string"
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "BlockchainNetworkServerStatus.StatusEnum"
        },
        {
            "name": "lastCheck",
            "baseName": "lastCheck",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return BlockchainNetworkServerStatus.attributeTypeMap;
    }
}

export namespace BlockchainNetworkServerStatus {
    export enum StatusEnum {
        healthy = <any> 'healthy',
        unhealthy = <any> 'unhealthy'
    }
}