// model/blockchainStatus.ts
/**
 * Key Connect API
 * Blockchain access made simple
 *
 * The version of the OpenAPI document: 1.1.0
 * Contact: manthan@keyconnect.app
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { BlockchainNetwork } from './blockchainNetwork';

/**
* Represents a blockchain and its status.
*/
export class BlockchainStatus {
    'chainId'?: string;
    'networks'?: Array<BlockchainNetwork>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "chainId",
            "baseName": "chainId",
            "type": "string"
        },
        {
            "name": "networks",
            "baseName": "networks",
            "type": "Array<BlockchainNetwork>"
        }    ];

    static getAttributeTypeMap() {
        return BlockchainStatus.attributeTypeMap;
    }
}