// model/rate.ts
/**
 * Key Connect API
 * Blockchain access made simple
 *
 * The version of the OpenAPI document: 1.1.0
 * Contact: manthan@keyconnect.app
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class Rate {
    /**
    * Base currency
    */
    'base'?: string;
    /**
    * Counter currency
    */
    'counter'?: string;
    /**
    * Rate to convert the base currency into the counter currency
    */
    'price'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "base",
            "baseName": "base",
            "type": "string"
        },
        {
            "name": "counter",
            "baseName": "counter",
            "type": "string"
        },
        {
            "name": "price",
            "baseName": "price",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return Rate.attributeTypeMap;
    }
}